+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Avoid using namespace alias and use isomorphic namespace in JSX runtime ([#249](https://github.com/yhatt/jsx-slack/pull/249))
+
 ## v4.4.0 - 2021-11-17
 
 ### Removed